Fixing a bug which caused RoomConfigurations to be 'not-deletable' when beeing imported from configuration file.
The bug was encumbered by the position the 'deletion listeners' were placed. They were only placed on the room configuration, when adding a single new room.
When importing a config file, all rooms are set alltogether, so the listeners were not placed.
Fixed by adding the listeners everytime a room is inserted.